The indirect interrupt mechanism allows me to just send IACs to the i960 from the AVR instead of some goofy bit mask design.
What I can do on the AVR side is queue up the IACs as we go through.

What I don't know about is what happens if the processor is given a bogus IAC. The Kx manual states there is a state for it but honestly I'm not sure what that will do. A test will need to be performed instead