Move the firmware image to start at 0x0100'0000
This is to support both a wide implementation and the working "narrow" implementation.

The narrow implementation supports up to 4 megabytes of ram which is mirrored over and over throughout the 32-bit memory space. So 0x0000'0000 and 0x0100'0000 map to the same address.

With the wide implementation, ram starts at 0x0100'0000 so we can copy the image out of boot flash (0x0000'0000 to 0x00FF'FFFF) into RAM and work with it that way.

It also means that the same concepts apply, with respect to the narrow implementation, the mega2560 is responsible for installing the image off of the SDCard. For the wide implementation, the i960 is responsible for doing the transfer out of flash into ram as normal.