Define the branch jump table so that we can use the instruction cache as much as possible
The i960Sx processor does not have a data cache or onboard RAM to offset
the expense of accessing the main bus. So we embed branch instructions
into the jump tables. This allows us to eliminate many memory bus
accesses when executing common sequences of code. It will be in the
cache itself.